Fixed conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@66870 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seq/seq_loc_from_string.cpp
@@ -333,7 +333,7 @@ namespace {
         str = copy;
     }
 
-    unsigned int s_GetParenLen (string text)
+    size_t s_GetParenLen (string text)
     {
         string::size_type offset = 0;
         unsigned int paren_count;
